requirements.txt
GitPython
networkx
matplotlib
jupyter
scipy
pytest